took off types for casting

+++ dqn_tpu.py
@@ -88,7 +88,7 @@ class DQN:
 
         q_values = strategy.run(
             self.model.predict,
-            args=(([(tf.convert_to_tensor(states, dtype=tf.uint8)), ((tf.convert_to_tensor(actions, dtype=tf.uint8)))])),
+            args=([tf.convert_to_tensor(states), tf.convert_to_tensor(actions)]),
         )
 
         #q_values = self.model.predict([np.array(states), np.array(actions)])
